refactor(tui): move layout constants to tui/constants.ts
