+++ src/tui/constants.ts
@@ -1,0 +1,5 @@
+/** Default column count when stdout has no TTY dimensions. */
+export const DEFAULT_COLUMNS = 120;
+
+/** Fallback terminal width for layout calculations. */
+export const DEFAULT_TERMINAL_WIDTH = 96;
