+++ src/tui/styles.ts
@@ -1,9 +1,3 @@
-/** Default column count when stdout has no TTY dimensions. */
-export const DEFAULT_COLUMNS = 120;
-
-/** Fallback terminal width for layout calculations. */
-export const DEFAULT_TERMINAL_WIDTH = 96;
-
 const ESC = "\x1b[";
 
 export const ansi = {
